Refactor and consolidate server middleware structure
Moves all middleware implementations to a new packages/server/src/middleware/ directory, splitting them into logical modules (auth, security, rate-limit, validation) and providing a single consolidated export. Updates all imports throughout the codebase to use the new middleware location. The old shared middleware file is now a deprecated re-export for backward compatibility.
