+++ packages/server/src/api/index.ts
@@ -22,7 +22,7 @@ import {
   securityMiddleware,
   validateContentTypeMiddleware,
   createApiRateLimit,
-} from './shared/middleware';
+} from '../middleware';
 
 /**
  * Sets up Socket.io server for real-time messaging
@@ -192,7 +192,7 @@ export function createPluginRouteHandler(elizaOS: ElizaOS): express.RequestHandl
               if (!res.headersSent) {
                 const status =
                   (error instanceof Error && 'code' in error && error.code === 'ENOENT') ||
-                  (error instanceof Error && error.message?.includes('not found'))
+                    (error instanceof Error && error.message?.includes('not found'))
                     ? 404
                     : 500;
                 res.status(status).json({
@@ -242,7 +242,7 @@ export function createPluginRouteHandler(elizaOS: ElizaOS): express.RequestHandl
               if (!res.headersSent) {
                 const status =
                   (error instanceof Error && 'code' in error && error.code === 'ENOENT') ||
-                  (error instanceof Error && error.message?.includes('not found'))
+                    (error instanceof Error && error.message?.includes('not found'))
                     ? 404
                     : 500;
                 res.status(status).json({
