+++ packages/server/src/middleware/index.ts
@@ -1,0 +1,27 @@
+/**
+ * Consolidated middleware for the ElizaOS server
+ * All middleware is organized into logical modules for better maintainability
+ */
+
+// Authentication middleware
+export { apiKeyAuthMiddleware } from './auth';
+
+// Security middleware
+export { securityMiddleware } from './security';
+
+// Rate limiting middleware
+export {
+    createApiRateLimit,
+    createFileSystemRateLimit,
+    createUploadRateLimit,
+    createChannelValidationRateLimit,
+} from './rate-limit';
+
+// Validation middleware
+export {
+    agentExistsMiddleware,
+    validateUuidMiddleware,
+    validateChannelIdMiddleware,
+    validateContentTypeMiddleware,
+} from './validation';
+
